Moved IBFT payload classes to common
