+++ consensus/common/src/main/java/org/hyperledger/besu/consensus/common/bft/payload/Authored.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.consensus.ibft.payload;
+package org.hyperledger.besu.consensus.common.bft.payload;
 
 import org.hyperledger.besu.ethereum.core.Address;
