+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/messagewrappers/IbftMessage.java
@@ -15,10 +15,10 @@
 package org.hyperledger.besu.consensus.ibft.messagewrappers;
 
 import org.hyperledger.besu.consensus.common.bft.ConsensusRoundIdentifier;
-import org.hyperledger.besu.consensus.ibft.payload.Authored;
-import org.hyperledger.besu.consensus.ibft.payload.Payload;
-import org.hyperledger.besu.consensus.ibft.payload.RoundSpecific;
-import org.hyperledger.besu.consensus.ibft.payload.SignedData;
+import org.hyperledger.besu.consensus.common.bft.payload.Authored;
+import org.hyperledger.besu.consensus.common.bft.payload.Payload;
+import org.hyperledger.besu.consensus.common.bft.payload.RoundSpecific;
+import org.hyperledger.besu.consensus.common.bft.payload.SignedData;
 import org.hyperledger.besu.ethereum.core.Address;
 import org.hyperledger.besu.ethereum.rlp.BytesValueRLPOutput;
