+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/payload/PreparedCertificate.java
@@ -14,6 +14,7 @@
  */
 package org.hyperledger.besu.consensus.ibft.payload;
 
+import org.hyperledger.besu.consensus.common.bft.payload.SignedData;
 import org.hyperledger.besu.ethereum.rlp.RLPInput;
 import org.hyperledger.besu.ethereum.rlp.RLPOutput;
